+++ app/src/main/java/ca/pkay/rcloneexplorer/Services/DownloadService.java
@@ -1,16 +1,19 @@
 package ca.pkay.rcloneexplorer.Services;
 
+import android.annotation.SuppressLint;
 import android.app.IntentService;
 import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.os.Build;
 import android.support.annotation.Nullable;
 import android.support.v4.app.NotificationCompat;
 
 import java.util.List;
+import java.util.concurrent.ExecutionException;
 
 import ca.pkay.rcloneexplorer.BroadcastReceivers.DownloadCancelAction;
 import ca.pkay.rcloneexplorer.Items.FileItem;
@@ -20,14 +23,19 @@ import ca.pkay.rcloneexplorer.Rclone;
 
 public class DownloadService extends IntentService {
 
-    private final String CHANNEL_ID = "ca.pkay.rcexplorer.download_channel";
-    private final String CHANNEL_NAME = "Downloads";
     public static final String DOWNLOAD_LIST_ARG = "ca.pkay.rcexplorer.download_service.arg1";
     public static final String DOWNLOAD_PATH_ARG = "ca.pkay.rcexplorer.download_service.arg2";
     public static final String REMOTE_ARG = "ca.pkay.rcexplorer.download_service.arg3";
+    private final String CHANNEL_ID = "ca.pkay.rcexplorer.download_channel";
+    private final String CHANNEL_NAME = "Downloads";
+    private final int PERSISTENT_NOTIFICATION_ID = 167;
+    private final int FAILED_DOWNLOAD_NOTIFICATION_ID = 138;
+    private final int DOWNLOAD_FINISHED_NOTIFICATION_ID = 80;
+    private int numOfRunningProcesses;
+    private int numOfFinishedDownloads;
+    private int numOfFailedDownloads;
     private Rclone rclone;
     private List<Process> runningProcesses;
-    private Boolean aProcessFailed;
 
     /**
      * Creates an IntentService.  Invoked by your subclass's constructor.*
@@ -58,7 +66,7 @@ public class DownloadService extends IntentService {
                 .setContentIntent(pendingIntent)
                 .addAction(R.drawable.ic_cancel_download, getString(R.string.cancel), cancelPendingIntent);
 
-        startForeground(1, builder.build());
+        startForeground(PERSISTENT_NOTIFICATION_ID, builder.build());
 
         if (intent == null) {
             return;
@@ -69,42 +77,51 @@ public class DownloadService extends IntentService {
         final String remote = intent.getStringExtra(REMOTE_ARG);
 
         runningProcesses = rclone.downloadItems(remote, downloadList, downloadPath);
-
+        numOfRunningProcesses = runningProcesses.size();
+        numOfFinishedDownloads = 0;
+        numOfFailedDownloads = 0;
+        AsyncTask[] asyncTasks = new AsyncTask[numOfRunningProcesses];
+        int i = 0;
         for (Process process : runningProcesses) {
-            try {
-                process.waitFor();
-                if (process.exitValue() != 0) {
-                    aProcessFailed = true;
-                }
+            asyncTasks[i++] = new MonitorDownload().execute(process);
+        }
 
-            } catch (InterruptedException e) {
+        for (AsyncTask asyncTask : asyncTasks) {
+            try {
+                asyncTask.get();
+            } catch (InterruptedException | ExecutionException e) {
                 e.printStackTrace();
             }
         }
 
         stopForeground(true);
+    }
 
-        NotificationCompat.Builder builder1;
-        if (aProcessFailed) {
-            builder1 = new NotificationCompat.Builder(this, CHANNEL_ID)
-                    .setSmallIcon(android.R.drawable.stat_sys_warning)
-                    .setContentTitle(getString(R.string.download_cancelled))
-                    .setPriority(NotificationCompat.PRIORITY_DEFAULT);
+    private void showDownloadFinishedNotification(int numOfFinishedDownloads, int numOfTotalDownloads) {
+        String notificationText = numOfFinishedDownloads + " " + getString(R.string.out_of) + " " + numOfTotalDownloads + " " + getString(R.string.downloads_finished);
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
+                .setSmallIcon(android.R.drawable.stat_sys_download_done)
+                .setContentTitle(getString(R.string.download_complete))
+                .setContentText(notificationText)
+                .setPriority(NotificationCompat.PRIORITY_LOW);
 
-            NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-            if (notificationManager != null) {
-                notificationManager.notify(2, builder1.build());
-            }
-        } else {
-            builder1 = new NotificationCompat.Builder(this, CHANNEL_ID)
-                    .setSmallIcon(android.R.drawable.stat_sys_download_done)
-                    .setContentTitle(getString(R.string.download_complete))
-                    .setPriority(NotificationCompat.PRIORITY_DEFAULT);
+        NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        if (notificationManager != null) {
+            notificationManager.notify(DOWNLOAD_FINISHED_NOTIFICATION_ID, builder.build());
+        }
+    }
 
-            NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
-            if (notificationManager != null) {
-                notificationManager.notify(3, builder1.build());
-            }
+    private void showDownloadFailedNotification(int numOfFailedDownloads, int numOfTotalDownloads) {
+        String notificationText = numOfFailedDownloads + " " + getString(R.string.out_of) + " " + numOfTotalDownloads + " " + getString(R.string.downloads_failed);
+        NotificationCompat.Builder builder = new NotificationCompat.Builder(this, CHANNEL_ID)
+                .setSmallIcon(android.R.drawable.stat_sys_warning)
+                .setContentTitle(getString(R.string.download_failed))
+                .setContentText(notificationText)
+                .setPriority(NotificationCompat.PRIORITY_LOW);
+
+        NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        if (notificationManager != null) {
+            notificationManager.notify(FAILED_DOWNLOAD_NOTIFICATION_ID, builder.build());
         }
     }
 
@@ -126,6 +143,31 @@ public class DownloadService extends IntentService {
             NotificationManager notificationManager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
             if (notificationManager != null) {
                 notificationManager.createNotificationChannel(channel);
+            }
+        }
+    }
+
+    @SuppressLint("StaticFieldLeak")
+    private class MonitorDownload extends AsyncTask<Process, Void, Boolean> {
+
+        @Override
+        protected Boolean doInBackground(Process... processes) {
+            Process process = processes[0];
+            try {
+                process.waitFor();
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            return process.exitValue() == 0;
+        }
+
+        @Override
+        protected void onPostExecute(Boolean result) {
+            super.onPostExecute(result);
+            if (result) {
+                showDownloadFinishedNotification(++numOfFinishedDownloads, numOfRunningProcesses);
+            } else {
+                showDownloadFailedNotification(++numOfFailedDownloads, numOfRunningProcesses);
             }
         }
     }